refactor: rename to scripts/extract-release-label.mjs
